add sort-cmp.cpp

+++ 04_search_sort/sort-cmp/sort-cmp.cpp
@@ -12,7 +12,7 @@ int partition(int arr[], int lo, int hi);
 int *swap(int arr[], int what, int with);
 
 int main() {
-  long long int help[] = {
+  long int help[] = {
     10,
     100,
     1000,
